健康检查任务优化
